+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+  "editor.tokenColorCustomizations": {
+    "comments": "",
+    "textMateRules": []
+  }
+}
