--- client/app.tsx
@@ -1,12 +1,0 @@
-// import React from 'react';
-// import { ThemeProvider } from 'react-native-elements';
-// import './config/firebase';
-// import RootNavigation from './nav';
-
-// export default function App() {
-//   return (
-//     <ThemeProvider>
-//       <RootNavigation />
-//     </ThemeProvider>
-//   );
-// }
