+++ client/app/_layout.tsx
@@ -1,14 +1,27 @@
-import React from 'react';
-import { Stack } from 'expo-router';
+// // import React from 'react';
+// // import { Stack } from 'expo-router';
 
 
-export default function TabLayout() {
-  return (
-    <Stack>
-      <Stack.Screen name="index" options={{ title: 'Home', headerShown: false }} />
-      <Stack.Screen name="PropertyDetails" options={{ title: 'Details'}} />
-      <Stack.Screen name="RequestaTour" options={{ title: 'Tour Request'}} />
-   
-    </Stack>
-  );
-}
+// // export default function TabLayout() {
+// //   return (
+// //     <Stack>
+// //       <Stack.Screen name="signIn" options={{ title: 'Home', headerShown: false }} />
+// //       <Stack.Screen name="PropertyDetails" options={{ title: 'Details' }} />
+// //       <Stack.Screen name="RequestaTour" options={{ title: 'Tour Request' }} />
+
+// //     </Stack>
+// //   );
+// // }
+
+// import React from 'react';
+// import { Stack } from 'expo-router';
+
+// export default function TabLayout() {
+//   return (
+//     <Stack>
+//       <Stack.Screen name="index" options={{ title: 'Welcome', headerShown: false }} />
+//       <Stack.Screen name="signIn" options={{ title: 'Sign In' }} />
+//       <Stack.Screen name="signUp" options={{ title: 'Sign Up' }} />
+//     </Stack>
+//   );
+// }
