+++ client/nav/userStack.tsx
@@ -2,7 +2,7 @@
 // import { NavigationContainer } from '@react-navigation/native';
 // import { createStackNavigator } from '@react-navigation/stack';
 
-// import HomeScreen from '../app/index';
+// import HomeScreen from '../app/home';
 
 // const Stack = createStackNavigator();
